requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake==6.3.0
apache-airflow-providers-cncf-kubernetes==10.1.0
# 